+++ src/containerapp/azext_containerapp/tests/latest/test_containerapp_commands.py
@@ -738,6 +738,10 @@ class ContainerappServiceBindingTests(ScenarioTest):
         self.cmd('containerapp service milvus create -g {} -n {} --environment {}'.format(
             resource_group, milvus_ca_name, env_name))
 
+        self.cmd('containerapp show -g {} -n {}'.format(resource_group, milvus_ca_name), checks=[
+            JMESPathCheck("properties.provisioningState", "Succeeded")
+        ])
+
         self.cmd('containerapp create -g {} -n {} --environment {} --image {} --bind postgres:postgres_binding redis'.format(
             resource_group, ca_name, env_name, image), checks=[
             JMESPathCheck('properties.template.serviceBinds[0].name', "postgres_binding"),
@@ -751,6 +755,7 @@ class ContainerappServiceBindingTests(ScenarioTest):
 
         self.cmd('containerapp update -g {} -n {} --bind postgres:postgres_binding kafka mariadb qdrant milvus'.format(
             resource_group, ca_name, image), checks=[
+            JMESPathCheck("properties.provisioningState", "Succeeded"),
             JMESPathCheck('properties.template.serviceBinds[0].name', "redis"),
             JMESPathCheck('properties.template.serviceBinds[1].name', "postgres_binding"),
             JMESPathCheck('properties.template.serviceBinds[2].name', "kafka"),