+++ src/containerapp/HISTORY.rst
@@ -3,8 +3,8 @@
 Release History
 ===============
 upcoming
-* 'az containerapp service': add support for creation and deletion of milvus vector database as a dev service
 ++++++
+* 'az containerapp service': add support for creation and deletion of milvus vector database as a dev service
 * 'az containerapp env create': Support --enable-dedicated-gpu
 * 'az containerapp job create': fix problem of parsing parameters minExecutions and maxExecutions from --yaml
 * 'az containerapp env dapr-component init': support initializing Dapr components and dev services for an environment